fix(test): keep public network tests out of PR CI
