+++ zebrad/tests/common/get_block_template_rpcs.rs
@@ -1,4 +1,4 @@
-//! Acceptance tests for getblocktemplate RPC methods in Zebra.
+//! Shared helpers for getblocktemplate RPC tests in Zebra.
 
 pub(crate) mod get_block_template;
 pub(crate) mod get_peer_info;
