+++ zebrad/tests/e2e/mod.rs
@@ -1,4 +1,6 @@
+pub mod rpc;
 pub mod sync;
+pub mod trusted_chain;
 
 #[cfg(feature = "zebra-checkpoints")]
 pub mod checkpoints;
