+++ zebrad/tests/e2e/sync.rs
@@ -10,6 +10,7 @@ use crate::common::sync::{
     create_cached_database_height, sync_until, MempoolBehavior, LARGE_CHECKPOINT_TEST_HEIGHT,
     LARGE_CHECKPOINT_TIMEOUT, MEDIUM_CHECKPOINT_TEST_HEIGHT, STOP_AT_HEIGHT_REGEX,
     STOP_ON_LOAD_TIMEOUT, SYNC_FINISHED_REGEX, TINY_CHECKPOINT_TEST_HEIGHT,
+    TINY_CHECKPOINT_TIMEOUT,
 };
 
 /// Test if `zebrad` can sync some larger checkpoints on mainnet.
@@ -60,6 +61,25 @@ fn sync_large_checkpoints_mempool_mainnet() -> Result<()> {
         &Mainnet,
         STOP_AT_HEIGHT_REGEX,
         LARGE_CHECKPOINT_TIMEOUT,
+        None,
+        MempoolBehavior::ForceActivationAt(TINY_CHECKPOINT_TEST_HEIGHT),
+        // checkpoint sync is irrelevant here - all tested checkpoints are mandatory
+        true,
+        true,
+    )
+    .map(|_tempdir| ())
+}
+
+/// Test if `zebrad` can activate the mempool on mainnet.
+/// Debug activation happens after committing the genesis block.
+#[test]
+#[ignore]
+fn activate_mempool_mainnet() -> Result<()> {
+    sync_until(
+        block::Height(TINY_CHECKPOINT_TEST_HEIGHT.0 + 1),
+        &Mainnet,
+        STOP_AT_HEIGHT_REGEX,
+        TINY_CHECKPOINT_TIMEOUT,
         None,
         MempoolBehavior::ForceActivationAt(TINY_CHECKPOINT_TEST_HEIGHT),
         // checkpoint sync is irrelevant here - all tested checkpoints are mandatory
