+++ zebrad/tests/integration/mod.rs
@@ -1,5 +1,4 @@
 pub mod database;
-pub mod mempool;
 #[cfg(not(target_os = "windows"))]
 pub mod network;
 pub mod regtest;
