+++ zebrad/tests/stateful/lightwalletd.rs
@@ -4,6 +4,20 @@ use crate::common::test_type::TestType::*;
 
 use crate::common::lightwalletd::lwd_integration_test;
 
+/// Make sure `lightwalletd` works with Zebra, when both their states are empty.
+///
+/// This test only runs when the `TEST_LIGHTWALLETD` env var is set.
+///
+/// This test doesn't work on Windows, so it is always skipped on that platform.
+#[test]
+#[ignore]
+#[cfg(not(target_os = "windows"))]
+fn lwd_integration() -> Result<()> {
+    lwd_integration_test(LaunchWithEmptyState {
+        launches_lightwalletd: true,
+    })
+}
+
 /// Make sure `lightwalletd` can sync from Zebra, in update sync mode.
 ///
 /// This test only runs when:
@@ -15,6 +29,7 @@ use crate::common::lightwalletd::lwd_integration_test;
 #[test]
 #[ignore]
 #[cfg(not(target_os = "windows"))]
+#[cfg(feature = "lightwalletd-grpc-tests")]
 fn lwd_sync_update() -> Result<()> {
     lwd_integration_test(UpdateCachedState)
 }
@@ -36,6 +51,7 @@ fn lwd_sync_update() -> Result<()> {
 #[tokio::test]
 #[ignore]
 #[cfg(not(target_os = "windows"))]
+#[cfg(feature = "lightwalletd-grpc-tests")]
 async fn lightwalletd_test_suite() -> Result<()> {
     lwd_integration_test(LaunchWithEmptyState {
         launches_lightwalletd: true,
@@ -44,28 +60,24 @@ async fn lightwalletd_test_suite() -> Result<()> {
     // Only runs when a cached Zebra state directory path is configured with an environment variable.
     lwd_integration_test(UpdateZebraCachedStateNoRpc)?;
 
-    // These tests need the compile-time gRPC feature
-    #[cfg(feature = "lightwalletd-grpc-tests")]
-    {
-        // Do the quick tests first
+    // Do the quick tests first
 
-        // Only runs when a cached Zebra state is configured
-        lwd_integration_test(UpdateCachedState)?;
+    // Only runs when a cached Zebra state is configured
+    lwd_integration_test(UpdateCachedState)?;
 
-        // Only runs when a cached Zebra state is configured
-        crate::common::lightwalletd::wallet_grpc_test::run().await?;
+    // Only runs when a cached Zebra state is configured
+    crate::common::lightwalletd::wallet_grpc_test::run().await?;
 
-        // Then do the slow tests
+    // Then do the slow tests
 
-        // Only runs when a cached Zebra state is configured.
-        // When manually running the test suite, allow cached state in the full sync test.
-        lwd_integration_test(FullSyncFromGenesis {
-            allow_lightwalletd_cached_state: true,
-        })?;
+    // Only runs when a cached Zebra state is configured.
+    // When manually running the test suite, allow cached state in the full sync test.
+    lwd_integration_test(FullSyncFromGenesis {
+        allow_lightwalletd_cached_state: true,
+    })?;
 
-        // Only runs when a cached Zebra state is configured
-        crate::common::lightwalletd::send_transaction_test::run().await?;
-    }
+    // Only runs when a cached Zebra state is configured
+    crate::common::lightwalletd::send_transaction_test::run().await?;
 
     Ok(())
 }
@@ -78,6 +90,7 @@ async fn lightwalletd_test_suite() -> Result<()> {
 #[tokio::test]
 #[ignore]
 #[cfg(not(target_os = "windows"))]
+#[cfg(feature = "lightwalletd-grpc-tests")]
 async fn lwd_rpc_send_tx() -> Result<()> {
     crate::common::lightwalletd::send_transaction_test::run().await
 }
@@ -90,6 +103,7 @@ async fn lwd_rpc_send_tx() -> Result<()> {
 #[tokio::test]
 #[ignore]
 #[cfg(not(target_os = "windows"))]
+#[cfg(feature = "lightwalletd-grpc-tests")]
 async fn lwd_grpc_wallet() -> Result<()> {
     crate::common::lightwalletd::wallet_grpc_test::run().await
 }
