+++ zebrad/tests/stateful/mod.rs
@@ -1,6 +1,5 @@
 #[cfg(feature = "indexer")]
 pub mod indexer;
-#[cfg(feature = "lightwalletd-grpc-tests")]
 pub mod lightwalletd;
 pub mod rpc;
 pub mod sync;
